MAB  Adding update_medicaid.cwl workflow to process already loaded raw data
Former-commit-id: 354e94fe0c563857fd73c02dc5a193e1a11dd9c5

+++ src/python/nsaph/loader/data_loader.py
@@ -120,10 +120,11 @@ class DataLoader(LoaderBase):
         if not self.context.reset:
             return
         with self._connect() as connxn:
-            tables = self.domain.drop(self.table, connxn)
+            if not self.context.sloppy:
+                tables = self.domain.drop(self.table, connxn)
+                for t in tables:
+                    IndexBuilder.drop_all(connxn, self.domain.schema, t)
             self.domain.create(connxn, [self.table])
-            for t in tables:
-                IndexBuilder.drop_all(connxn, self.domain.schema, t)
 
     def drop(self):
         schema = self.domain.schema
